+++ packages/client/base/src/types/embed/v3/events/incoming.ts
@@ -4,12 +4,16 @@ export const embeddedCheckoutV3IncomingEvents = {
     "ui:height.changed": z.object({
         height: z.number(),
     }),
+    "crypto:load": z.object({}),
     "crypto:connect-wallet.show": z.object({
         show: z.boolean(),
     }),
     "crypto:send-transaction": z.object({
         chain: z.string(),
         serializedTransaction: z.string(),
+    }),
+    "crypto:sign-message": z.object({
+        message: z.string(),
     }),
     "order:updated": z.object({
         order: z.any().optional(),