+++ packages/wallets/src/signers/shadow-signer/evm-shadow-signer.ts
@@ -1,24 +1,23 @@
 import { ShadowSigner } from "./shadow-signer";
 import type { EVMSmartWalletChain } from "@/chains/chains";
-import type { EVM256KeypairInternalSignerConfig } from "../types";
+import type { P256KeypairInternalSignerConfig } from "../types";
 import type { ShadowSignerData } from "./utils";
-import { EVM256KeypairSigner } from "../evm-p256-keypair";
+import { P256KeypairSigner } from "../p256-keypair";
 
 export class EVMShadowSigner extends ShadowSigner<
     EVMSmartWalletChain,
-    EVM256KeypairSigner,
-    EVM256KeypairInternalSignerConfig
+    P256KeypairSigner,
+    P256KeypairInternalSignerConfig
 > {
     protected getSignerClass() {
-        return EVM256KeypairSigner;
+        return P256KeypairSigner;
     }
 
-    getShadowSignerConfig(shadowData: ShadowSignerData): EVM256KeypairInternalSignerConfig {
+    getShadowSignerConfig(shadowData: ShadowSignerData): P256KeypairInternalSignerConfig {
         return {
-            type: "evm-p256-keypair",
-            publicKey: shadowData.publicKeyBase64,
-            chain: shadowData.chain as EVMSmartWalletChain,
-            locator: `evm-p256-keypair:${shadowData.chain}:${shadowData.publicKeyBase64}`,
+            type: "p256-keypair",
+            address: shadowData.publicKeyBase64,
+            locator: `p256-keypair:${shadowData.publicKeyBase64}`,
             onSignTransaction: async (pubKey: string, data: Uint8Array) => {
                 return await this.storage.sign(pubKey, data);
             },